added member struct to factory

+++ src/abstracts/MintingStages.sol
@@ -23,7 +23,7 @@ abstract contract MintingStages is Clone, AccessControl, ERC721A, IERC2981 {
     Stages public mintingStages;
     Collection public collectionData;
     address internal _platformFeeReceiver;
-    uint96 internal _platformFee;
+    uint256 internal _platformFee;
     bool internal _initalized = false;
 
     event UpdateWLevent(address indexed sender, uint256 listLength);
